working ubo

+++ src/plugins/lighting/blinn_phong/shaders/vert.glsl
@@ -16,7 +16,7 @@ layout (std140) uniform lighting_blinn_phong_UBO {
     vec3 u_ambient_color;
     float u_ambient_intensity;
     
-    float u_nlights;
+    int u_nlights;
     
     float u_light_intensities[MAX_LIGHTS_N];
     float u_light_specular_exp[MAX_LIGHTS_N];
@@ -28,27 +28,11 @@ out vec4 normal;
 out vec3 light_dirs[MAX_LIGHTS_N];
 out vec3 light_half_vects[MAX_LIGHTS_N];
 
-//<TOBE REMOVED>
-// vec3 ulp[3];
-//</TOBE REMOVED>
-
 int compute_lighting_vert(mat4 view_m, mat4 modelview_m) {
     vec4 view_pos = modelview_m * vec4(a_position, 1.0);
     
-    //<TOBE REMOVED>
-    // ulp[0] = vec3(60.5, 10., 50.);
-    // ulp[1] = vec3(-1.5, -0.0, 10.);
-    // ulp[2] = vec3(-50.5, 10., 10.);
-    //</TOBE REMOVED>
-    
     for(int i = 0; i < int(u_nlights); i ++ ) {
-        //<TOBE ADDED>
         vec4 lpos = view_m * vec4(u_light_positions[i], 1.0);
-        //</TOBE ADDED>
-        
-        //<TOBE REMOVED>
-        // vec4 lpos = view_m * vec4(ulp[i], 1.0);
-        //</TOBE REMOVED>
         
         light_dirs[i] = normalize(lpos.xyz - view_pos.xyz);
